==> notebooks/01_iob_dataset_preparation.ipynb <==
import os, glob, json, spacy, pickle

import pandas as pd
import numpy as np

from functools import reduce

nlp = spacy.load('pt_core_news_lg')

annotated = glob.glob(r'../data/raw/v2/*.json')

annotated

jfile = ' '.join(open(annotated[1], 'r', encoding='utf-8').readlines())

TAG_TRANSLATE = {
    '"DESPESAPUBLICA"' : '"CONTROLEEXTERNO"',
    '"DESPESAPESSOAL"' : '"CONTROLEEXTERNO"',
    '"CONTRATO"' : '"CONTROLEEXTERNO"',
    '"PROCESSO"' : '"CONTROLEEXTERNO"',
    '"CPF"' : '"DOCUMENTO"',
    '"CNPJ"' : '"DOCUMENTO"',
    '"LICITACAO"' : '"CONTROLEEXTERNO"',
    '"PF"' : '"PESSOA"',
    '"PJ"' : '"ORGANIZACAO"',
    '"ORGAO"' : '"ORGANIZACAO"'    
}

final_json = []
for f in annotated:
    c = ' '.join(open(f, 'r', encoding='utf-8').readlines())
    for k,v in TAG_TRANSLATE.items():
        c = c.replace(k,v)
    final_json.extend(json.loads(c))

with open(r'../data/raw/v2/annotated.pkl', 'wb') as f:
    pickle.dump(final_json, f)
