==> notebooks/02_template_generation.ipynb <==
import os, glob, json, spacy, pickle
import random

class TemplateGenerator():
    def __init__(self):
        self.templates = json.load(open(r'../data/interim/templates.json', 'r', encoding='utf-8'))
        entidades = json.load(open(r'../data/interim/ents_para_template.json', 'r', encoding='latin9'))
        with open(r'../data/interim/despesapessoal.txt', 'r', encoding='utf-8') as f:
            self.despesa_pessoal = set([x.replace('\n','') for x in f.readlines() if x])
        with open(r'../data/interim/despesapublica.txt', 'r', encoding='utf-8') as f:
            self.despesa_publica = set([x.replace('\n','') for x in f.readlines() if x])
        with open(r'../data/interim/substantivos.txt', 'r', encoding='latin9') as f:
            self.substantivos = list(set([x.replace('\n','') for x in f.readlines() if x]))
            self.substantivos = list(filter(lambda x:x not in self.despesa_pessoal and x not in self.despesa_publica, 
                                       self.substantivos))
        
        self.utterances = {}
        self.utterances['DESPESAPESSOAL'] = []
        self.utterances['DESPESAPUBLICA'] = []
        self.utterances['GENERICA'] = []
        
    def generate(self):
        # pessoal
        self.tpessoal = self.templates['DESPESAPESSOAL']
        for t in self.tpessoal:
            for d in self.despesa_pessoal:
                self.utterances['DESPESAPESSOAL'].append(t.replace('{ENTIDADE}',d))
                
        self.tpublica = self.templates['DESPESAPUBLICA']
        for t in self.tpublica:
            for d in self.despesa_publica:
                self.utterances['DESPESAPUBLICA'].append(t.replace('{ENTIDADE}',d))
                
        len_generica = max(len(self.utterances['DESPESAPUBLICA']), len(self.utterances['DESPESAPESSOAL']))
        tgen = self.tpessoal + self.tpublica
        for _ in range(len_generica):
            t = random.choice(tgen)
            s = random.choice(self.substantivos)
            self.utterances['GENERICA'].append(t.replace('{ENTIDADE}',s))
            
    def save(self):
        with open(r'../data/interim/utterances.json', 'w', encoding='utf-8') as fp:
            fp.write(json.dumps(self.utterances, indent=2, ensure_ascii=False))
            
    def save_train_data(self):
        self.train_data = []
        
        self.tpessoal = self.templates['DESPESAPESSOAL']
        for t in self.tpessoal:
            for d in self.despesa_pessoal:
                cnt = t.replace('{ENTIDADE}',d)
                edict = {'entities':[(cnt.find(d),cnt.find(d)+len(d),"CONTROLEEXTERNO")]}
                self.train_data.append((cnt, edict))
                
                
        self.tpublica = self.templates['DESPESAPUBLICA']
        for t in self.tpublica:
            for d in self.despesa_publica:
                cnt = t.replace('{ENTIDADE}',d)
                edict = {'entities':[(cnt.find(d),cnt.find(d)+len(d),"CONTROLEEXTERNO")]}
                self.train_data.append((cnt, edict))
        
        with open(r'../data/interim/synth_train_data.pkl', 'wb') as fp:
            pickle.dump(self.train_data, fp)
       
        

t = TemplateGenerator()

t.generate()

t.save_train_data()

t.save()

Geração do embedding específico

with open(r'data\gazetteers\despesapessoal.txt', 'r', encoding='utf-8') as f:
    despesa_pessoal = set([x.replace('\n','') for x in f.readlines() if x])
with open(r'data\gazetteers\despesapublica.txt', 'r', encoding='utf-8') as f:
    despesa_publica = set([x.replace('\n','') for x in f.readlines() if x])

gazetteers = list(despesa_pessoal) + list(despesa_publica)

nlp

corpus = [[str(x) for x in nlp(g)] for g in gazetteers]

model = Word2Vec(corpus, min_count=1, size=30, workers=3, window=2, sg=1)

model.wv['siai-dp']

model.wv.save_word2vec_format(r'data\wordemb\specific.txt')
