==> notebooks/03_dataset_preparation.ipynb <==
import os, glob, json, spacy, re
import pandas as pd
import numpy as np

from functools import reduce

from spacy.util import compile_prefix_regex, compile_infix_regex, compile_suffix_regex
from gensim.models import KeyedVectors

Carregamento dos modelos de embedding

genmodel = KeyedVectors.load_word2vec_format(r'data\wordemb\fasttext50.txt')

spemodel = KeyedVectors.load_word2vec_format(r'data\wordemb\specific.txt')

nlp = spacy.load("pt_core_news_lg")

Arquivos de texto para treinamento

annotated = glob.glob(r'data\annotated\*.json')

all_content = []
for j in annotated:
    jfile = json.load(open(j, 'r', encoding='utf-8'))
    for f in jfile:
        all_content.append(f['content'])

from functools import reduce
all_words = reduce(lambda x,y: x+y, [[t.text for t in nlp(c)] for c in all_content])

with open(r"data\trainset\utterances.json", "r", encoding='utf-8') as fp:
    jtrainset = json.load(fp)

trainset = pd.DataFrame(columns=['tipo','frase'])

trainset['frase_len'] = trainset.frase.apply(lambda x: len(nlp(x)))

MAX_LEN = trainset['frase_len'].max()

Character Embedding

import keras
from keras.preprocessing.text import one_hot,Tokenizer
from keras.preprocessing.sequence import pad_sequences
from keras.models import Sequential
from keras.layers import Dense, Flatten, Embedding, Input
from keras.models import Model

CHAR_VOCAB = r"abcdefghijklmnopqrstuvwxyzçàáéíóúãõâêîôûABCDEFGHIJKLMNOPQRSTUVWXYZÇÀÁÉÍÓÚÃÕÂÊÎÔÛ\
,\"`'“”:;?. !()[]{}<>+–-*/=<>.#$%&@/\^_|~%§°ºª1234567890•"
CHAR_DICT = {v:i+1 for i,v in enumerate(CHAR_VOCAB)}
CHAR_DICT['unk'] = 0
MAX_CHAR_LEN = 50

def word2number(c):
    try:
        return CHAR_DICT[c]
    except:
        return CHAR_DICT['unk']

words_dataset = pad_sequences([[word2number(c) for c in w] for w in all_words], maxlen=50, padding='post')

l = list(CHAR_DICT.values())

VOCAB_SIZE = max(l)+1

def custom_tokenizer(nlp):
    infix_re = re.compile(r'''[\,\?\:\;\‘\’\`\“\”\"~]''')
    prefix_re = compile_prefix_regex(nlp.Defaults.prefixes)
    suffix_re = compile_suffix_regex(nlp.Defaults.suffixes)

    return Tokenizer(nlp.vocab, prefix_search=prefix_re.search,
                                suffix_search=suffix_re.search,
                                infix_finditer=infix_re.finditer,
                                token_match=None)

def sentence_to_tokens(s):
    nlp.tokenizer = custom_tokenizer(nlp)
    doc = nlp(s)
    return [t.text if t.text not in ['\n','\t'] else '' for t in doc]

class CharCNN():
    def __init__(self, vocab, dataset):
        self.vocab = vocab
        self.vocab_size = len(vocab)
        self.ds = dataset
        
    def build_model(self):
        word_input = Input(shape=(1,), dtype='float64')
        word_embedding = Embedding(input_dim=VOCAB_SIZE,
                                   output_dim=1,
                                   input_length=1)(word_input)
        word_vec = Flatten()(word_embedding)
        self.model = Model([word_input],word_vec)
        self.model.compile(optimizer=keras.optimizers.Adam(lr=1e-3),loss='binary_crossentropy',metrics=['acc'])
        #print(self.model.summary())
    
    
    def get_embedding(self, sent):
        words = [t.text for t in nlp(sent)]
        seqs = pad_sequences([[word2number(c) for c in w] for w in words], maxlen=50, padding='post')
        if self.model:
            return [self.model.predict(s).flatten() for s in seqs]
        else:
            return seqs
    
    def load_model(self):
        pass
    
    def save_model(self):
        pass

cnn = CharCNN(CHAR_VOCAB, words_dataset)

cnn.build_model()

cnn.get_embedding("oi tudo bem")

cnn.model.predict(words_dataset[0]).flatten().shape

(50*50 + 50 + 30)

Transformação em vetores

class VectorTransformer():
    def __init__(self, trainset, max_sent_len):
        self.ds = trainset
        self.max_sent_len = max_sent_len
        self.vecs = []
        self.nlp = spacy.load("pt_core_news_lg")
        
        self.genemb = KeyedVectors.load_word2vec_format(r'data\wordemb\fasttext50.txt')
        self.speemb = KeyedVectors.load_word2vec_format(r'data\wordemb\specific.txt')
        
        self.genemb.vocab['<pad>'] = np.zeros((self.genemb.vector_size,))        
        self.speemb.vocab['<pad>'] = np.zeros((self.speemb.vector_size,))
        
        self.charemb = CharCNN(CHAR_VOCAB, words_dataset)
        self.charemb.build_model()
        
    def get_vec(self, sent):
        words = [str(a) for a in nlp(sent)]
        if len(words) < self.max_sent_len:
            words += ['<pad>'] * (self.max_sent_len - len(words))
        elif len(words) > self.max_sent_len:
            words = words[:len(words)]
            
        vecs = []
        for w in words:
            wrd_vec = []
            
            try:
                wrd_vec.append(genmodel[w])
            except:
                wrd_vec.append(np.zeros((genmodel.vector_size,)))
            
            try:
                wrd_vec.append(spemodel[w])
            except:
                wrd_vec.append(np.zeros((spemodel.vector_size,)))
                
            wrd_vec.append(self.charemb.get_embedding(w)[0])
            
            vecs.append(reduce(lambda x,y: np.append(x,y), wrd_vec))
                
        return vecs
        
    def set2vec(self):
        for i,r in self.ds.iterrows():
            self.vecs.append((r['tipo'], np.array(self.get_vec(r['frase']))))
            
    

v = VectorTransformer(trainset, 23)

v.set2vec()

v.vecs[0][1].shape

v.get_vec("oi tudo bem")

Salvando para rede neural

CLASS_DICT = {'GENERICA':0,'DESPESAPESSOAL':1,'DESPESAPUBLICA':2}

vectors = [(CLASS_DICT[v[0]],v[1]) for v in v.vecs]

df_dataset = pd.DataFrame(columns=['label','features'], data=vectors)

df_dataset.to_pickle(r'data\trainset\dataset.pkl')
