==> notebooks/04_neural_network.ipynb <==
import os, glob, json

import numpy as np
import pandas as pd

import keras

from keras.models import Sequential, Model
from keras.layers import LSTM, Input
from keras.layers import Dense
from keras.layers import TimeDistributed
from keras.layers import Bidirectional
from keras.utils import to_categorical

from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report, confusion_matrix

Conjunto de teste

dataset = pd.read_pickle(r'../data/trainset/dataset.pkl')

X = np.array([[list(y) for y in list(x)] for x in dataset.features.values])
#X = dataset.features.values
y = dataset.label.values
y = to_categorical(y)

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.33, random_state=42)

X.shape

y[0]

y_test

INPUT_SHAPE = (23,130)

Modelos

class BiLTSM():
    def __init__(self, n_hidden_layers, units_per_layer, dropout=0.5):
        self.n_hidden_layers = n_hidden_layers
        self.units_per_layer = units_per_layer
        self.dropout = dropout
    
    def build_model(self):
        metrcs = ['accuracy']
        '''
        self.model = Sequential();
        for i in range(self.n_hidden_layers-1):
            self.model.add(Bidirectional(LSTM(units=self.units_per_layer,
                        input_shape=INPUT_SHAPE,
                      return_sequences=False)))
        self.model.add(Bidirectional(LSTM(self.units_per_layer,
                        input_shape=INPUT_SHAPE,
                      return_sequences=False)))
        self.model.add(Dense(3, activation='softmax'))
        '''
        inp = Input(shape=(23, 130))
        x = Bidirectional(LSTM(120, return_sequences=True))(inp)
        x = Bidirectional(LSTM(120, return_sequences=True))(x)
        x = Bidirectional(LSTM(120, return_sequences=False))(x)
        x = Dense(3, activation='softmax')(x)
        
        self.model = Model(inp, x)

        #self.model.summary()
        self.model.compile(loss='categorical_crossentropy',optimizer=keras.optimizers.Adam(), metrics=metrcs)
        
    def fit_model(self, epochs):
        self.model.fit(X_train, y_train, 
                       validation_split=.3, 
                       epochs=epochs, batch_size=32)

model1 = BiLTSM(1, 120)

model1.build_model()

model1.model.summary()

model1.fit_model(100)

y_test = np.argmax(y_test, axis=1) # Convert one-hot to index
y_pred = model1.model.predict(X_test, verbose=1)
y_pred = np.argmax(y_pred,axis=1)

print(classification_report(y_test, y_pred))

print(confusion_matrix(y_test, y_pred))
